feat: like everything (#60)
* minor optimizations (#54)

* feat: more optimizations akin to previous PR (#55)

* minor optimizations

* optimize

* optimize

* optimize

* fix: various optimizations (#56)

* optimize simulator

I should note that while this *technically* optimizes the actual app as well, but the Swift compiler for on-device will notice isSimulator will never be set, so compiler optimizations will remove it and, in the end, produce the same code as this optimization. However, since it is set on Simulator, it won't optimize it, meaning this produces slower code. I know Simulator optimizations don't matter to an average user, but this change (slightly) benefits development (and even if it won't matter for the on-device app, doing this optimization in the code itself rather than relying on compiler optimizations to do it for you is a better practice :P).

* optimize simulator

I should note that while this *technically* optimizes the actual app as well, but the Swift compiler for on-device will notice isSimulator will never be set, so compiler optimizations will remove it and, in the end, produce the same code as this optimization. However, since it is set on Simulator, it won't optimize it, meaning this produces slower code. I know Simulator optimizations don't matter to an average user, but this change (slightly) benefits development (and even if it won't matter for the on-device app, doing this optimization in the code itself rather than relying on compiler optimizations to do it for you is a better practice :P).

* remove IOKit

At least from what I can see, there is no place where the loader uses anything from it(?). I tried compiling without it, and everything seems fine and doesn't look to have any effect on the compiled binary.

Removing this makes development experience much better, as before you couldn't compile this on systems without this IOKit header, now with the changes it's possible.

* minor optimization

* minor optimization: we already know it's rootless

* optimize JsonVC

Improved logic. No longer preforms a lookup table when section is 0. When section is 1, instead of doing the lookup table we replace it entirely with if statements since two comparisons are faster than a lookup table.

* more optimization

* comment changes for clarity

* spawn is always called as root

since spawn is always called as root, we can remove this arg

* spawn is always called as root

since spawn is always called as root, we can remove this arg

* spawn is always called as root

since spawn is always called as root, we can remove this arg

* spawn is always called as root

since spawn is always called as root, we can remove this arg

* spawn is always called as root

since spawn is always called as root, we can remove this arg

* remove systemVersion from envInfo

this is a slight optimization

* remove systemVersion from envInfo

this is a slight optimization

* very minor optimization: just use envInfo.jsonURI itself

* feat: use palera1n dir instead because organization ig

* chore: FIX PLEASE

* feat: loader iOS 12.2 support

I made this support the lowest version it can really go without it needing to bundle any swift libs, so if anyone wants to use this loader they can even if they're not on a supported iOS version of palera1n, however in the helper theres some hardcoded values so when bootstrapping using palera1ns helper it would require iOS 15, but if you manage to make it support lower good on you !!

* feat: New Crowdin updates (#58)

* feat(lang): Update Romanian translations

* feat(lang): Update French translations

* feat(lang): Update Spanish translations

* feat(lang): Update Afrikaans translations

* feat(lang): Update Arabic translations

* feat(lang): Update Catalan translations

* feat(lang): Update Czech translations

* feat(lang): Update Danish translations

* feat(lang): Update German translations

* feat(lang): Update Greek translations

* feat(lang): Update Finnish translations

* feat(lang): Update Hebrew translations

* feat(lang): Update Hungarian translations

* feat(lang): Update Italian translations

* feat(lang): Update Japanese translations

* feat(lang): Update Korean translations

* feat(lang): Update Lithuanian translations

* feat(lang): Update Dutch translations

* feat(lang): Update Norwegian translations

* feat(lang): Update Polish translations

* feat(lang): Update Portuguese translations

* feat(lang): Update Russian translations

* feat(lang): Update Slovak translations

* feat(lang): Update Albanian translations

* feat(lang): Update Serbian (Cyrillic) translations

* feat(lang): Update Swedish translations

* feat(lang): Update Turkish translations

* feat(lang): Update Ukrainian translations

* feat(lang): Update Chinese Simplified translations

* feat(lang): Update Chinese Traditional translations

* feat(lang): Update English translations

* feat(lang): Update Vietnamese translations

* feat(lang): Update Portuguese, Brazilian translations

* feat(lang): Update Indonesian translations

* feat(lang): Update Persian translations

* feat(lang): Update Pirate English translations

* feat(lang): Update English (upside down) translations

* feat(lang): Update LOLCAT translations

* feat: dev stuff

* feat: BETTER bootstrap handling because apple

* chore: move stuff around

* feat: use assets finally (loader side)

* feat: change install check & clean stuff up

* feat: unneeded stuff

* chore: update readme

* chore: fix pale/kern info flag parsing

* feat: mwah

* Create JSON-DOCS.md

* Update JSON-DOCS.md

* ass code

* a

---------

Co-authored-by: 0xilis <109871561+0xilis@users.noreply.github.com>
Co-authored-by: staturnz <118011249+staturnzz@users.noreply.github.com>